requirements.txt
streamlit==1.47.1
pandas==2.3.1
requests==2.32.4
mysql-connector==2.2.9
pytest==8.4.1
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0